Remove NetResolver.cpp/h (almost empty)
