+++ CMakeLists.txt
@@ -2389,8 +2389,6 @@ add_library(${CoreLibName} ${CoreLinkType}
 	Core/MIPS/MIPSAsm.h
 	Core/MIPS/MIPSTracer.cpp
 	Core/MIPS/MIPSTracer.h
-	Core/Net/NetResolver.cpp
-	Core/Net/NetResolver.h
 	Core/MemFault.cpp
 	Core/MemFault.h
 	Core/MemMap.cpp
