+++ Core/Core.vcxproj.filters
@@ -94,9 +94,6 @@
     <Filter Include="MIPS\RiscV">
       <UniqueIdentifier>{067e3128-3aaf-4ed1-b19e-bab11606abe7}</UniqueIdentifier>
     </Filter>
-    <Filter Include="Net">
-      <UniqueIdentifier>{f1b3f688-26eb-4121-9914-e9e19ab04bcd}</UniqueIdentifier>
-    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="ELF\ElfReader.cpp">
@@ -1333,9 +1330,6 @@
     <ClCompile Include="HLE\sceNetResolver.cpp">
       <Filter>HLE\Libraries</Filter>
     </ClCompile>
-    <ClCompile Include="Net\NetResolver.cpp">
-      <Filter>Net</Filter>
-    </ClCompile>
     <ClCompile Include="HLE\NetInetConstants.cpp">
       <Filter>HLE\Libraries</Filter>
     </ClCompile>
@@ -2150,9 +2144,6 @@
     </ClInclude>
     <ClInclude Include="HLE\sceNetResolver.h">
       <Filter>HLE\Libraries</Filter>
-    </ClInclude>
-    <ClInclude Include="Net\NetResolver.h">
-      <Filter>Net</Filter>
     </ClInclude>
     <ClInclude Include="HLE\NetInetConstants.h">
       <Filter>HLE\Libraries</Filter>
