--- Core/Net/NetResolver.cpp
@@ -1,3 +1,0 @@
-#include "NetResolver.h"
-
-// TODO: remove me if the functions remain inlined
